+++ src/channel.rs
@@ -6,42 +6,42 @@ mod upsampler;
 use crate::audio_utils::to_dbfs;
 
 pub struct Channel {
-        rms: f64,
-        peak: f64,
+        rms: f32,
+        peak: f32,
         clip_sample_count: i32,
         true_clip_sample_count: i32,
-        dc_offset: f64,
+        dc_offset: f32,
         samples_count: i32,
         upsampled_samples_count: i32,
-        true_peak: f64,
+        true_peak: f32,
 }
 
 impl Channel {
-        pub fn rms(&self) -> f64 {
+        pub fn rms(&self) -> f32 {
                 to_dbfs(self.rms)
         }
 
-        pub fn peak(&self) -> f64 {
+        pub fn peak(&self) -> f32 {
                 to_dbfs(self.peak)
         }
 
-        pub fn true_peak(&self) -> f64 {
+        pub fn true_peak(&self) -> f32 {
                 to_dbfs(self.true_peak)
         }
 
-        pub fn clip_samples_quota(&self) -> f64 {
-                (self.clip_sample_count as f64 / self.samples_count as f64) * 100.0
+        pub fn clip_samples_quota(&self) -> f32 {
+                (self.clip_sample_count as f32 / self.samples_count as f32) * 100.0
         }
 
-        pub fn true_clip_samples_quota(&self) -> f64 {
-                (self.true_clip_sample_count as f64 / self.upsampled_samples_count as f64) * 100.0
+        pub fn true_clip_samples_quota(&self) -> f32 {
+                (self.true_clip_sample_count as f32 / self.upsampled_samples_count as f32) * 100.0
         }
 
-        pub fn dc_offset(&self) -> f64 {
+        pub fn dc_offset(&self) -> f32 {
                 self.dc_offset
         }
 
-        pub fn crest_factor(&self) -> f64 {
+        pub fn crest_factor(&self) -> f32 {
                 to_dbfs(self.peak / self.rms)
         }
